feat: split signal page into Events/Settings tabs
Drop the cards row + jobs/runs tabs from the signal page. Settings tab
embeds the manage-signal form directly via a new ManageSignalPanel
variant; the existing sheet remains for the create-signal flow on the
signals list. Test panel becomes a slide-out drawer in panel mode.
