+++ frontend/components/signals/create-signal-drawer/enum-values-input.tsx
