+++ frontend/components/signals/create-signal-drawer/sampling-section.tsx
