+++ frontend/components/signals/create-signal-drawer/schema-fields-builder.tsx
