+++ frontend/components/signals/create-signal-drawer/template-picker/index.tsx
