+++ frontend/components/signals/create-signal-drawer/template-picker/template-item.tsx
