+++ frontend/components/signals/create-signal-drawer/test-panel/payload-value.tsx
