+++ frontend/components/signals/create-signal-drawer/test-panel/test-results-view.tsx
