+++ frontend/components/signals/create-signal-drawer/triggers-section.tsx
